+++ sdv/metadata/dataset.py
@@ -10,6 +10,7 @@ import numpy as np
 import pandas as pd
 from rdt import HyperTransformer, transformers
 
+from sdv.constraints import Constraint
 from sdv.metadata import visualization
 from sdv.metadata.errors import MetadataError
 
@@ -477,7 +478,7 @@ class Metadata:
         dtypes = self.get_dtypes(table_name)
         pii_fields = self._get_pii_fields(table_name)
         transformers_dict = self._get_transformers(dtypes, pii_fields)
-        return HyperTransformer(transformers=transformers_dict)
+        return HyperTransformer(field_transformers=transformers_dict)
 
     def transform(self, table_name, data):
         """Transform data for a given table.
@@ -873,7 +874,7 @@ class Metadata:
         return fields_metadata
 
     def add_table(self, name, data=None, fields=None, fields_metadata=None,
-                  primary_key=None, parent=None, foreign_key=None):
+                  primary_key=None, parent=None, foreign_key=None, constraints=None):
         """Add a new table to this metadata.
 
         ``fields`` list can be a mixture of field names, which will be build automatically
@@ -904,7 +905,10 @@ class Metadata:
             parent (str):
                 Table name to refere a foreign key field. Defaults to ``None``.
             foreign_key (str):
-                Foreing key field name to ``parent`` table primary key. Defaults to ``None``.
+                Foreign key field name to ``parent`` table primary key. Defaults to ``None``.
+            constraints (list[Constraint, dict]):
+                List of Constraint objects or dicts representing the constraints for the
+                given table.
 
         Raises:
             ValueError:
@@ -939,6 +943,16 @@ class Metadata:
             table_metadata['path'] = path
 
         self._metadata['tables'][name] = table_metadata
+
+        if constraints:
+            meta_constraints = []
+            for constraint in constraints:
+                if isinstance(constraint, Constraint):
+                    meta_constraints.append(constraint.to_dict())
+                else:
+                    meta_constraints.append(constraint)
+
+            table_metadata['constraints'] = meta_constraints
 
         try:
             if primary_key: