+++ tests/unit/metadata/test_table.py
@@ -2,6 +2,8 @@ from unittest.mock import Mock, patch
 
 import pandas as pd
 import pytest
+from faker import Faker
+from faker.config import DEFAULT_LOCALE
 from rdt.transformers.numerical import NumericalTransformer
 
 from sdv.constraints.base import Constraint
@@ -11,6 +13,223 @@ from sdv.metadata import Table
 
 
 class TestTable:
+
+    def test__get_faker_default_locale(self):
+        """Test that ``_get_faker`` without locales parameter has default locale.
+
+        The ``_get_faker`` should return a Faker object localized to the default locale.
+        When no locales are specified explicitly.
+
+        Input:
+        - Field metadata from metadata dict.
+        Output:
+        - Faker object with default localization.
+        """
+        # Setup
+        metadata_dict = {
+            'fields': {
+                'foo': {
+                    'type': 'categorical',
+                    'pii': True,
+                    'pii_category': 'company'
+                }
+            }
+        }
+
+        # Run
+        faker = Table.from_dict(metadata_dict)._get_faker(metadata_dict['fields']['foo'])
+
+        # Assert
+        assert isinstance(faker, Faker)
+        assert faker.locales == [DEFAULT_LOCALE]
+
+    def test__get_faker_specified_locales_string(self):
+        """Test that ``_get_faker`` with locales parameter sets localization correctly.
+
+        The ``_get_faker`` should return a Faker object localized to the specified locale.
+
+        Input:
+        - Field metadata from metadata dict.
+        Output:
+        - Faker object with specified localization string.
+        """
+        # Setup
+        metadata_dict = {
+            'fields': {
+                'foo': {
+                    'type': 'categorical',
+                    'pii': True,
+                    'pii_category': 'company',
+                    'pii_locales': 'sv_SE'
+                }
+            }
+        }
+
+        # Run
+        faker = Table.from_dict(metadata_dict)._get_faker(metadata_dict['fields']['foo'])
+
+        # Assert
+        assert isinstance(faker, Faker)
+        assert faker.locales == ['sv_SE']
+
+    def test__get_faker_specified_locales_list(self):
+        """Test that ``_get_faker`` with locales parameter sets localization correctly.
+
+        The ``_get_faker`` should return a Faker object localized to the specified locales.
+
+        Input:
+        - Field metadata from metadata dict.
+        Output:
+        - Faker object with specified list of localizations.
+        """
+        # Setup
+        metadata_dict = {
+            'fields': {
+                'foo': {
+                    'type': 'categorical',
+                    'pii': True,
+                    'pii_category': 'company',
+                    'pii_locales': ['en_US', 'sv_SE']
+                }
+            }
+        }
+
+        # Run
+        faker = Table.from_dict(metadata_dict)._get_faker(metadata_dict['fields']['foo'])
+
+        # Assert
+        assert isinstance(faker, Faker)
+        assert faker.locales == ['en_US', 'sv_SE']
+
+    def test__get_faker_method_pass_args(self):
+        """Test that ``_get_faker_method`` method utilizes parameters passed in category argument.
+
+        The ``_get_faker_method`` method uses the parameters passed to it in the category argument.
+
+        Input:
+        - Faker object to create faked values with.
+        - Category tuple of category name and parameters passed to the method creating fake values.
+        Output:
+        - Fake values created with the specified method from the Faker object.
+        Utilizing the arguments given to it.
+        """
+        # Setup
+        metadata_dict = {
+            'fields': {
+                'foo': {
+                    'type': 'categorical',
+                    'pii': True,
+                    'pii_category': 'ean'
+                }
+            }
+        }
+        metadata = Table.from_dict(metadata_dict)
+
+        # Run
+        fake_8_ean = metadata._get_faker_method(Faker(), ('ean', 8))
+        ean_8 = fake_8_ean()
+
+        fake_13_ean = metadata._get_faker_method(Faker(), ('ean', 13))
+        ean_13 = fake_13_ean()
+
+        # Assert
+        assert len(ean_8) == 8
+        assert len(ean_13) == 13
+
+    @patch('sdv.metadata.Table')
+    def test__make_anonymization_mappings(self, mock_table):
+        """Test that ``_make_anonymization_mappings`` creates the expected mappings.
+
+        The ``_make_anonymization_mappings`` method should map values in the original
+        data to fake values for non-id fields that are labeled pii.
+
+        Setup:
+        - Create a Table that has metadata about three fields (one pii field, one id field,
+          and one non-pii field).
+        Input:
+        - Data that contains a pii field, an id field, and a non-pii field.
+        Side Effects:
+        - Expect ``_get_fake_values`` to be called with the number of unique values of the
+          pii field.
+        - Expect the resulting `_ANONYMIZATION_MAPPINGS` field to contain the pii field, with
+          the correct number of mappings and keys.
+        """
+        # Setup
+        metadata = Mock()
+        metadata._ANONYMIZATION_MAPPINGS = {}
+        foo_metadata = {
+            'type': 'categorical',
+            'pii': True,
+            'pii_category': 'email',
+        }
+        metadata._fields_metadata = {
+            'foo': foo_metadata,
+            'bar': {
+                'type': 'categorical',
+            },
+            'baz': {
+                'type': 'id',
+            }
+        }
+        foo_values = ['test1@example.com', 'test2@example.com', 'test3@example.com']
+        data = pd.DataFrame({
+            'foo': foo_values,
+            'bar': ['a', 'b', 'c'],
+            'baz': [1, 2, 3],
+        })
+
+        # Run
+        Table._make_anonymization_mappings(metadata, data)
+
+        # Assert
+        assert mock_table._get_fake_values.called_once_with(foo_metadata, 3)
+
+        mappings = metadata._ANONYMIZATION_MAPPINGS[id(metadata)]
+        assert len(mappings) == 1
+
+        foo_mappings = mappings['foo']
+        assert len(foo_mappings) == 3
+        assert list(foo_mappings.keys()) == foo_values
+
+    @patch('sdv.metadata.Table')
+    def test__make_anonymization_mappings_unique_faked_value_in_field(self, mock_table):
+        """Test that ``_make_anonymization_mappings`` method creates mappings for anonymized values.
+
+        The ``_make_anonymization_mappings`` method should map equal values in the original data
+        to the same faked value.
+
+        Input:
+        - DataFrame with a field that should be anonymized based on the metadata description.
+        Side Effect:
+        - Mappings are created from the original values to faked values.
+        """
+        # Setup
+        metadata = Mock()
+        metadata._ANONYMIZATION_MAPPINGS = {}
+        foo_metadata = {
+            'type': 'categorical',
+            'pii': True,
+            'pii_category': 'email'
+        }
+        metadata._fields_metadata = {
+            'foo': foo_metadata
+        }
+        data = pd.DataFrame({
+            'foo': ['test1@example.com', 'test2@example.com', 'test1@example.com']
+        })
+
+        # Run
+        Table._make_anonymization_mappings(metadata, data)
+
+        # Assert
+        assert mock_table._get_fake_values.called_once_with(foo_metadata, 2)
+
+        mappings = metadata._ANONYMIZATION_MAPPINGS[id(metadata)]
+        assert len(mappings) == 1
+
+        foo_mappings = mappings['foo']
+        assert len(foo_mappings) == 2
+        assert list(foo_mappings.keys()) == ['test1@example.com', 'test2@example.com']
 
     @patch.object(Constraint, 'from_dict')
     def test__prepare_constraints_sorts_constraints(self, from_dict_mock):